# backend/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Beatris API
  version: 1.0.0
  description: API веб-приложения медицинского центра эстетической медицины Beatris
servers:
  - url: /api/v1
    description: Основной относительный адрес API
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Зарегистрировать пользователя
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          headers:
            Set-Cookie:
              description: HttpOnly cookie с refresh token
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Авторизовать пользователя
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Пользователь успешно авторизован
          headers:
            Set-Cookie:
              description: HttpOnly cookie с refresh token
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Обновить access token
      description: Использует refresh token из защищённой HttpOnly cookie и возвращает новый access token.
      operationId: refreshAccessToken
      security:
        - refreshCookie: []
      responses:
        "200":
          description: Access token успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Завершить сеанс пользователя
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Пользователь успешно вышел из системы
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/me:
    get:
      tags:
        - Auth
      summary: Получить текущего пользователя
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные текущего пользователя успешно получены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /appointments/available-slots:
    get:
      tags:
        - Appointments
      summary: Получить свободное время для записи
      operationId: getAvailableSlots
      parameters:
        - name: specialistId
          in: query
          description: Идентификатор специалиста
          required: true
          schema:
            type: string
            format: uuid
        - name: serviceId
          in: query
          description: Идентификатор услуги
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          description: Дата для поиска свободных слотов
          required: true
          schema:
            type: string
            format: date
            example: "2026-05-20"
      responses:
        "200":
          description: Свободное время успешно получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableSlot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments:
    post:
      tags:
        - Appointments
      summary: Создать запись или заявку на приём
      description: Endpoint доступен гостю и авторизованному пациенту. Если пользователь авторизован, данные пациента могут быть подставлены из профиля.
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Запись на приём успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/my:
    get:
      tags:
        - Appointments
      summary: Получить мои записи
      operationId: getMyAppointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Записи пациента успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/{id}/cancel:
    patch:
      tags:
        - Appointments
      summary: Отменить запись на приём
      operationId: cancelAppointment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Запись успешно отменена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /cart:
    get:
      tags:
        - Cart
      summary: Получить корзину пользователя
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина успешно получена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cart/items:
    post:
      tags:
        - Cart
      summary: Добавить товар в корзину
      operationId: addCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "201":
          description: Товар успешно добавлен в корзину
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cart/items/{itemId}:
    patch:
      tags:
        - Cart
      summary: Изменить количество товара в корзине
      operationId: updateCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          description: Идентификатор позиции корзины
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Количество товара успешно изменено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Cart
      summary: Удалить товар из корзины
      operationId: deleteCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          description: Идентификатор позиции корзины
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Товар успешно удалён из корзины
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token, передаваемый в заголовке Authorization
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Refresh token, хранящийся в защищённой HttpOnly cookie
  parameters:
    Id:
      name: id
      in: path
      description: Уникальный идентификатор ресурса
      required: true
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 400
            message: Некорректные данные запроса
            error: Bad Request
    Unauthorized:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 401
            message: Необходимо выполнить вход в систему
            error: Unauthorized
    Forbidden:
      description: Недостаточно прав доступа
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 403
            message: Недостаточно прав для выполнения действия
            error: Forbidden
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 404
            message: Запрашиваемый ресурс не найден
            error: Not Found
    ValidationError:
      description: Ошибка валидации данных
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            statusCode: 422
            message:
              - phone должен быть указан
              - serviceId должен быть UUID
            error: Unprocessable Entity
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 500
            message: Внутренняя ошибка сервера
            error: Internal Server Error
  schemas:
    Role:
      type: string
      enum:
        - patient
        - doctor
        - admin
      example: patient
    AppointmentStatus:
      type: string
      enum:
        - pending
        - confirmed
        - completed
        - cancelled
      example: confirmed
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Некорректные данные запроса
        error:
          type: string
          example: Bad Request
    ValidationErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
          example: 422
        message:
          type: array
          items:
            type: string
          example:
            - phone должен быть указан
            - serviceId должен быть UUID
        error:
          type: string
          example: Unprocessable Entity
    User:
      type: object
      description: Пользователь системы
      required:
        - id
        - fullName
        - phone
        - role
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        fullName:
          type: string
          example: Айгуль Нурланова
        phone:
          type: string
          example: "+77001234567"
        email:
          type: string
          format: email
          nullable: true
          example: user@example.com
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-20T10:00:00Z"
    RegisterRequest:
      type: object
      required:
        - fullName
        - phone
        - password
      properties:
        fullName:
          type: string
          example: Айгуль Нурланова
        phone:
          type: string
          example: "+77001234567"
        email:
          type: string
          format: email
          nullable: true
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: qwerty123
    LoginRequest:
      type: object
      required:
        - phone
        - password
      properties:
        phone:
          type: string
          example: "+77001234567"
        password:
          type: string
          format: password
          example: qwerty123
    AuthResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Время жизни access token в секундах
          example: 900
        user:
          $ref: "#/components/schemas/User"
    RefreshTokenResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 900
...
    Service:
      type: object
      description: Услуга медицинского центра
      required:
        - id
        - title
        - slug
        - description
        - price
        - category
      properties:
        id:
          type: string
          format: uuid
          example: 661b9091-86a4-4512-aa0d-162c8d0c2db0
        title:
          type: string
          example: Биоревитализация
        slug:
          type: string
          example: biorevitalizatsiya
        description:
          type: string
          example: Инъекционная процедура для увлажнения и восстановления кожи
        price:
          type: number
          minimum: 0
          example: 25000
        durationMinutes:
          type: integer
          nullable: true
          example: 60
        category:
          $ref: "#/components/schemas/ServiceCategory"
        imageUrl:
          type: string
          nullable: true
          example: /uploads/services/biorevitalizatsiya.jpg
        contraindications:
          type: string
          nullable: true
          example: Беременность, воспалительные процессы, индивидуальные противопоказания
...
    Specialist:
      type: object
      description: Специалист медицинского центра
      required:
        - id
        - fullName
        - position
        - specialization
      properties:
        id:
          type: string
          format: uuid
          example: 4bd38fe7-6d07-43c6-81c1-47e7e1279356
        fullName:
          type: string
          example: Иванова Анна Сергеевна
        position:
          type: string
          example: Врач-косметолог
        specialization:
          type: string
          example: Инъекционная косметология
        experienceYears:
          type: integer
          example: 8
        photoUrl:
          type: string
          nullable: true
          example: /uploads/specialists/ivanova.jpg
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
...
    AvailableSlot:
      type: object
      required:
        - specialistId
        - date
        - time
      properties:
        specialistId:
          type: string
          format: uuid
        date:
          type: string
          format: date
          example: "2026-05-20"
        time:
          type: string
          example: "14:30"
    CreateAppointmentRequest:
      type: object
      required:
        - serviceId
        - specialistId
        - date
        - time
        - patientName
        - patientPhone
      properties:
        serviceId:
          type: string
          format: uuid
        specialistId:
          type: string
          format: uuid
        date:
          type: string
          format: date
          example: "2026-05-20"
        time:
          type: string
          example: "14:30"
        patientName:
          type: string
          example: Айгуль Нурланова
        patientPhone:
          type: string
          example: "+77001234567"
        comment:
          type: string
          nullable: true
          example: Хочу записаться на первичную консультацию
    Appointment:
      type: object
      description: Запись на приём
      required:
        - id
        - service
        - specialist
        - date
        - time
        - status
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
          nullable: true
        patientName:
          type: string
          example: Айгуль Нурланова
        patientPhone:
          type: string
          example: "+77001234567"
        service:
          $ref: "#/components/schemas/Service"
        specialist:
          $ref: "#/components/schemas/Specialist"
        date:
          type: string
          format: date
          example: "2026-05-20"
        time:
          type: string
          example: "14:30"
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        comment:
          type: string
          nullable: true
          example: Первичная консультация
        patientContactStatus:
          type: string
          enum:
            - not_contacted
            - contacted
            - agreed
            - declined
          example: agreed
        patientContactedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-09T10:00:00Z"
        patientContactComment:
          type: string
          nullable: true
          example: Пациент согласовал перенос на 15:30
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentStatusHistoryItem"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
...
    AppointmentStatusHistoryItem:
      type: object
      required:
        - id
        - previousStatus
        - newStatus
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        adminId:
          type: string
          format: uuid
          nullable: true
        adminName:
          type: string
          nullable: true
          example: Администратор Беатрис
        previousStatus:
          $ref: "#/components/schemas/AppointmentStatus"
        newStatus:
          $ref: "#/components/schemas/AppointmentStatus"
        createdAt:
          type: string
          format: date-time
          example: "2026-06-09T10:00:00Z"
...
    AddCartItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          example: 1
    UpdateCartItemRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1
          example: 2
    CartItem:
      type: object
      required:
        - id
        - product
        - quantity
        - price
        - subtotal
      properties:
        id:
          type: string
          format: uuid
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
          example: 2
        price:
          type: number
          example: 18000
        subtotal:
          type: number
          example: 36000
    Cart:
      type: object
      required:
        - items
        - totalPrice
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalPrice:
          type: number
          example: 36000
...